+++ spec/install/gems/post_install_spec.rb
@@ -46,39 +46,6 @@ describe "bundle install with gem sources" do
   end
 end
 
-describe "bundle install with path sources" do
-  describe "when gems include post install messages" do
-    it "should display the post-install messages after installing" do
-      build_lib "foo", :path => lib_path('foo') do |s|
-        s.post_install_message = "Foo's post install message"
-      end
-      gemfile <<-G
-        source "file://#{gem_repo1}"
-        gem 'foo', :path => '#{lib_path("foo")}'
-      G
-
-      bundle :install
-      expect(out).to include("Post-install message from foo:")
-      expect(out).to include("Foo's post install message")
-    end
-  end
-
-  describe "when gems do not include post install messages" do
-    it "should not display any post-install messages" do
-      build_lib "foo", :path => lib_path('foo') do |s|
-        s.post_install_message = nil
-      end
-      gemfile <<-G
-        source "file://#{gem_repo1}"
-        gem 'foo', :path => '#{lib_path("foo")}'
-      G
-
-      bundle :install
-      expect(out).not_to include("Post-install message")
-    end
-  end
-end
-
 describe "bundle install with git sources" do
   describe "when gems include post install messages" do
     it "should display the post-install messages after installing" do